Phase 4 Step 1: Fix Jest configuration and add lobby API tests
- Fix Jest/Babel configuration conflicts that were breaking tests
- Change 'import type' to 'import' in lobby files to resolve Babel parsing issues
- Add MockScene export to setupTests.js for test compatibility
- Fix TypeScript type issues in test files (InstanceType, optional chaining)
- Add comprehensive API client test for lobby functionality
- Update lobby store to use real API endpoints instead of mock data
- All 244 tests now passing across 20 test suites

Resolves Jest configuration issues and establishes working test foundation
for Phase 4 frontend integration work.
