--- .babelrc
@@ -1,1 +1,0 @@
-{"presets":[["@babel/preset-env",{"targets":{"node":"current"}}],"@babel/preset-typescript"]}
