+++ jest.config.js
@@ -2,55 +2,82 @@
 module.exports = {
   verbose: true,
   testTimeout: 10000,
-  maxWorkers: 1,  // Force serial execution globally to prevent database deadlocks
-  transform: {
-    '^.+\\.(ts|tsx)$': [
-      'ts-jest',
-      {
-        tsconfig: 'tsconfig.json',
-        diagnostics: {
-          ignoreCodes: [1343]
-        },
-        isolatedModules: false,
-        useESM: false,
-        astTransformers: {
-          before: [
-            {
-              path: 'ts-jest/dist/transformers/hoist-jest',
-              options: { enums: true }
-            }
-          ]
-        }
-      }
-    ]
-  },
+  maxWorkers: 1, // prevent DB deadlocks
+
+  // Shared defaults (projects can override)
+  transformIgnorePatterns: ['/node_modules/(?!(phaser)/)'],
   moduleNameMapper: {
     '^phaser$': '<rootDir>/node_modules/phaser/dist/phaser.js',
-    '^@/(.*)$': '<rootDir>/src/$1'
+    '^@/(.*)$': '<rootDir>/src/$1',
   },
-  transformIgnorePatterns: [
-    'node_modules/(?!(phaser)/)'
-  ],
   moduleDirectories: ['node_modules', 'src'],
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+
   projects: [
     {
       displayName: 'client',
       testEnvironment: 'jsdom',
       testMatch: ['<rootDir>/src/client/__tests__/**/*.test.ts'],
-      setupFilesAfterEnv: ['<rootDir>/src/client/__tests__/setupTests.ts']
+      setupFilesAfterEnv: ['<rootDir>/src/client/__tests__/setupTests.js'], // use the JS setup we just fixed
+      testEnvironmentOptions: { url: 'http://localhost' },
+
+      // IMPORTANT: put ts transform INSIDE the project
+      transform: {
+        '^.+\\.(ts|tsx)$': [
+          'ts-jest',
+          {
+            tsconfig: 'tsconfig.json',
+            diagnostics: { ignoreCodes: [1343] },
+            isolatedModules: false,
+            useESM: false,
+          },
+        ],
+      },
+      transformIgnorePatterns: ['/node_modules/(?!(phaser)/)'],
+      moduleNameMapper: {
+        '^phaser$': '<rootDir>/node_modules/phaser/dist/phaser.js',
+        '^@/(.*)$': '<rootDir>/src/$1',
+      },
     },
+
     {
       displayName: 'server',
       testEnvironment: 'node',
       testMatch: ['<rootDir>/src/server/__tests__/**/*.test.ts'],
       setupFilesAfterEnv: ['<rootDir>/src/server/__tests__/setup.ts'],
-      maxWorkers: 1  // Force serial execution for database tests to prevent deadlocks
+      maxWorkers: 1,
+
+      transform: {
+        '^.+\\.(ts|tsx)$': [
+          'ts-jest',
+          {
+            tsconfig: 'tsconfig.json',
+            diagnostics: { ignoreCodes: [1343] },
+            isolatedModules: false,
+            useESM: false,
+          },
+        ],
+      },
+      transformIgnorePatterns: ['/node_modules/'],
     },
+
     {
       displayName: 'shared',
       testEnvironment: 'node',
-      testMatch: ['<rootDir>/src/shared/**/__tests__/**/*.test.ts']
-    }
-  ]
-}; 
+      testMatch: ['<rootDir>/src/shared/**/__tests__/**/*.test.ts'],
+
+      transform: {
+        '^.+\\.(ts|tsx)$': [
+          'ts-jest',
+          {
+            tsconfig: 'tsconfig.json',
+            diagnostics: { ignoreCodes: [1343] },
+            isolatedModules: false,
+            useESM: false,
+          },
+        ],
+      },
+      transformIgnorePatterns: ['/node_modules/'],
+    },
+  ],
+};
