+++ src/client/lobby/store/lobby.store.ts
@@ -1,6 +1,6 @@
 // store/lobby.store.ts
 import { create } from 'zustand';
-import type { Game, Player, ApiError, CreateGameForm, JoinGameForm, ID } from '../shared/types';
+import { Game, Player, ApiError, CreateGameForm, JoinGameForm, ID } from '../shared/types';
 import { api } from '../shared/api';
 
 // Mock data for development
